feat: Double -> Float 타입 변경

+++ Android/app/src/main/java/com/example/myapplication/data/dto/request/measure/BloodPressureRequest.kt
@@ -1,6 +1,6 @@
 package com.example.myapplication.data.dto.request.measure
 
 data class BloodPressureRequest(
-    val dbp: Double,
-    val sbp: Double
+    val dbp: Float,
+    val sbp: Float
 )
