+++ Android/app/src/main/java/com/example/myapplication/data/dto/request/measure/BodyCompositionRequest.kt
@@ -1,14 +1,14 @@
 package com.example.myapplication.data.dto.request.measure
 
 data class BodyCompositionRequest(
-    val bfm: Double,
-    val bfp: Double,
-    val bmr: Double,
-    val bodyAge: Int,
-    val ecf: Double,
-    val ecw: Double,
-    val icw: Double,
-    val mineral: Double,
-    val protein: Double,
-    val smm: Double
+    val bfm: Float,
+    val bfp: Float,
+    val bmr: Float,
+    val bodyAge: Short,
+    val ecf: Float,
+    val ecw: Float,
+    val icw: Float,
+    val mineral: Float,
+    val protein: Float,
+    val smm: Float
 )
