+++ Android/app/src/main/java/com/example/myapplication/data/dto/request/measure/HeartRateRequest.kt
@@ -1,6 +1,6 @@
 package com.example.myapplication.data.dto.request.measure
 
 data class HeartRateRequest(
-    val bpm: Int,
-    val oxygen: Int
+    val bpm: Short,
+    val oxygen: Short
 )
