+++ Android/app/src/main/java/com/example/myapplication/data/dto/request/measure/StressRequest.kt
@@ -1,8 +1,8 @@
 package com.example.myapplication.data.dto.request.measure
 
 data class StressRequest(
-    val bpm: Int,
-    val oxygen: Int,
+    val bpm: Short,
+    val oxygen: Short,
     val stressLevel: String,
-    val stressValue: Int
+    val stressValue: Short
 )
