+++ Android/app/src/main/java/com/example/myapplication/data/dto/request/measure/TemperatureRequest.kt
@@ -1,5 +1,5 @@
 package com.example.myapplication.data.dto.request.measure
 
 data class TemperatureRequest(
-    val temperature: Double
+    val temperature: Float
 )
